Each job as its own ws log subscription and channel
Improves concurrency and allows management of
connection on a per job basis
