+++ services/notification_listener_test.go
@@ -1,8 +1,6 @@
 package services_test
 
 import (
-	"encoding/json"
-	"strings"
 	"testing"
 
 	"github.com/ethereum/go-ethereum/common"
@@ -14,7 +12,6 @@ import (
 	"github.com/smartcontractkit/chainlink/store/models"
 	"github.com/smartcontractkit/chainlink/utils"
 	"github.com/stretchr/testify/assert"
-	"github.com/tidwall/gjson"
 )
 
 func TestNotificationListener_Start_NewHeads(t *testing.T) {
@@ -60,7 +57,6 @@ func newAddr() common.Address {
 
 func TestNotificationListener_AddJob_Listening(t *testing.T) {
 	t.Parallel()
-
 	sharedAddr := newAddr()
 	noAddr := common.Address{}
 
@@ -73,10 +69,9 @@ func TestNotificationListener_AddJob_Listening(t *testing.T) {
 		data      hexutil.Bytes
 	}{
 		{"ethlog matching address", "ethlog", sharedAddr, sharedAddr, 1, hexutil.Bytes{}},
-		{"ethlog non-matching address", "ethlog", newAddr(), newAddr(), 0, hexutil.Bytes{}},
+		{"ethlog all address", "ethlog", noAddr, newAddr(), 1, hexutil.Bytes{}},
 		{"runlog w/o address", "runlog", noAddr, newAddr(), 1, cltest.StringToRunLogData(`{"value":"100"}`)},
 		{"runlog matching address", "runlog", sharedAddr, sharedAddr, 1, cltest.StringToRunLogData(`{"value":"100"}`)},
-		{"runlog non-matching", "runlog", newAddr(), newAddr(), 0, hexutil.Bytes{}},
 	}
 
 	for _, test := range tests {
@@ -120,45 +115,6 @@ func TestNotificationListener_AddJob_Listening(t *testing.T) {
 	}
 }
 
-func jsonFromFixture(path string) models.JSON {
-	res := gjson.Get(string(cltest.LoadJSON(path)), "params.result")
-	out := cltest.JSONFromString(res.String())
-	return out
-}
-
-func TestStore_FormatLogJSON(t *testing.T) {
-	t.Parallel()
-
-	var clData models.JSON
-	clDataFixture := `{"url":"https://etherprice.com/api","path":["recent","usd"],"address":"0x3cCad4715152693fE3BC4460591e3D3Fbd071b42","dataPrefix":"0x0000000000000000000000000000000000000000000000000000000000000001","functionSelector":"76005c26"}`
-	assert.Nil(t, json.Unmarshal([]byte(clDataFixture), &clData))
-
-	hwLog := cltest.LogFromFixture("../internal/fixtures/eth/subscription_logs_hello_world.json")
-	exampleLog := cltest.LogFromFixture("../internal/fixtures/eth/subscription_logs.json")
-	tests := []struct {
-		name        string
-		el          types.Log
-		initr       models.Initiator
-		wantErrored bool
-		wantData    models.JSON
-	}{
-		{"example ethLog", exampleLog, models.Initiator{Type: "ethlog"}, false,
-			jsonFromFixture("../internal/fixtures/eth/subscription_logs.json")},
-		{"hello world ethLog", hwLog, models.Initiator{Type: "ethlog"}, false,
-			jsonFromFixture("../internal/fixtures/eth/subscription_logs_hello_world.json")},
-		{"hello world runLog", hwLog, models.Initiator{Type: "runlog"}, false,
-			clData},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			output, err := services.FormatLogJSON(test.initr, test.el)
-			assert.JSONEq(t, strings.ToLower(test.wantData.String()), strings.ToLower(output.String()))
-			assert.Equal(t, test.wantErrored, (err != nil))
-		})
-	}
-}
-
 func TestNotificationListener_newHeadsNotification(t *testing.T) {
 	t.Parallel()
 
@@ -167,8 +123,7 @@ func TestNotificationListener_newHeadsNotification(t *testing.T) {
 	store := app.Store
 
 	ethMock := app.MockEthClient()
-	nhChan := make(chan models.BlockHeader)
-	ethMock.RegisterSubscription("newHeads", nhChan)
+	nhChan := ethMock.RegisterNewHeads()
 	ethMock.Register("eth_getTransactionReceipt", strpkg.TxReceipt{})
 	sentAt := uint64(23456)
 	confirmationAt := sentAt + 1
@@ -198,136 +153,5 @@ func TestNotificationListener_newHeadsNotification(t *testing.T) {
 	nhChan <- models.BlockHeader{blockNumber}
 
 	ethMock.EnsureAllCalled(t)
-	assert.Equal(t, blockNumber, app.NotificationListener.HeadTracker.Get().Number)
-}
-
-func TestServices_InitiatorsForLog(t *testing.T) {
-	t.Parallel()
-
-	store, cleanup := cltest.NewStore()
-	defer cleanup()
-
-	runLogSig := services.RunLogTopic
-	requestID := common.StringToHash("42")
-
-	elj := cltest.NewJob()
-	el := models.Initiator{Type: "ethlog", Address: cltest.NewAddress()}
-	elj.Initiators = []models.Initiator{el}
-	assert.Nil(t, store.SaveJob(&elj))
-	assert.Nil(t, store.One("JobID", elj.ID, &el))
-
-	rlj := cltest.NewJob()
-	rl := models.Initiator{Type: "runlog"}
-	rlj.Initiators = []models.Initiator{rl}
-	assert.Nil(t, store.SaveJob(&rlj))
-	assert.Nil(t, store.One("JobID", rlj.ID, &rl))
-	rljIDHash := common.StringToHash(rlj.ID)
-
-	rlaj := cltest.NewJob()
-	rla := models.Initiator{Type: "runlog", Address: cltest.NewAddress()}
-	rlaj.Initiators = []models.Initiator{rla}
-	assert.Nil(t, store.SaveJob(&rlaj))
-	assert.Nil(t, store.One("JobID", rlaj.ID, &rla))
-	rlajIDHash := common.StringToHash(rlaj.ID)
-
-	tests := []struct {
-		name string
-		log  types.Log
-		want []models.Initiator
-	}{
-		{"ethlog matching address", types.Log{Address: el.Address},
-			[]models.Initiator{el}},
-		{"ethlog non-matching address", types.Log{Address: cltest.NewAddress()},
-			[]models.Initiator{}},
-		{"runlog w/ required topic", types.Log{
-			Topics: []common.Hash{runLogSig, requestID, rljIDHash},
-		}, []models.Initiator{rl}},
-		{"runlog w/o required topic", types.Log{
-			Topics: []common.Hash{runLogSig, requestID, cltest.NewHash()},
-		}, []models.Initiator{}},
-		{"runlog w/ matching address", types.Log{
-			Address: rla.Address,
-			Topics:  []common.Hash{runLogSig, requestID, rlajIDHash},
-		}, []models.Initiator{rla}},
-		{"runlog w/o matching address", types.Log{
-			Address: cltest.NewAddress(),
-			Topics:  []common.Hash{runLogSig, requestID, rlajIDHash},
-		}, []models.Initiator{}},
-		{"runlog w/ matching address but not job ID", types.Log{
-			Address: cltest.NewAddress(),
-			Topics:  []common.Hash{runLogSig, requestID, cltest.NewHash()},
-		}, []models.Initiator{}},
-		{"runlog matching ethlog address", types.Log{
-			Address: el.Address,
-			Topics:  []common.Hash{runLogSig, requestID, rljIDHash},
-		}, []models.Initiator{el, rl}},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			actual, err := services.InitiatorsForLog(store, test.log)
-			assert.Nil(t, err)
-			assert.Equal(t, test.want, actual)
-		})
-	}
-}
-
-// If updating this test, be sure to update the truffle suite's "expected event signature" test.
-func TestServices_RunLogTopic_ExpectedEventSignature(t *testing.T) {
-	t.Parallel()
-
-	expected := "0x06f4bf36b4e011a5c499cef1113c2d166800ce4013f6c2509cab1a0e92b83fb2"
-	assert.Equal(t, expected, services.RunLogTopic.Hex())
-}
-
-func TestHeadTracker_New(t *testing.T) {
-	t.Parallel()
-
-	store, cleanup := cltest.NewStore()
-	defer cleanup()
-	assert.Nil(t, store.Save(&models.BlockHeader{cltest.BigHexInt(1)}))
-	last := models.BlockHeader{cltest.BigHexInt(10)}
-	assert.Nil(t, store.Save(&last))
-	assert.Nil(t, store.Save(&models.BlockHeader{cltest.BigHexInt(2)}))
-
-	ht, err := services.NewHeadTracker(store)
-	assert.Nil(t, err)
-	assert.Equal(t, last.Number, ht.Get().Number)
-}
-
-func TestHeadTracker_Get(t *testing.T) {
-	t.Parallel()
-
-	store, cleanup := cltest.NewStore()
-	defer cleanup()
-	initial := models.BlockHeader{cltest.BigHexInt(1)}
-	assert.Nil(t, store.Save(&initial))
-
-	tests := []struct {
-		name      string
-		toSave    *models.BlockHeader
-		want      hexutil.Big
-		wantError bool
-	}{
-		// order matters
-		{"greater", &models.BlockHeader{cltest.BigHexInt(2)}, cltest.BigHexInt(2), false},
-		{"less than", &models.BlockHeader{cltest.BigHexInt(1)}, cltest.BigHexInt(2), false},
-		{"zero", &models.BlockHeader{cltest.BigHexInt(0)}, cltest.BigHexInt(2), true},
-		{"nil", nil, cltest.BigHexInt(2), true},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			ht, err := services.NewHeadTracker(store)
-			assert.Nil(t, err)
-			err = ht.Save(test.toSave)
-			if test.wantError {
-				assert.NotNil(t, err)
-			} else {
-				assert.Nil(t, err)
-			}
-
-			assert.Equal(t, test.want, ht.Get().Number)
-		})
-	}
+	assert.Equal(t, blockNumber, app.Store.HeadTracker.Get().Number)
 }
