+++ store/store.go
@@ -1,11 +1,14 @@
 package store
 
 import (
+	"errors"
 	"os"
 	"os/signal"
+	"sync"
 	"syscall"
 	"time"
 
+	"github.com/asdine/storm"
 	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/smartcontractkit/chainlink/logger"
 	"github.com/smartcontractkit/chainlink/store/models"
@@ -15,12 +18,13 @@ import (
 // for keeping the application state in sync with the database.
 type Store struct {
 	*models.ORM
-	Config    Config
-	Clock     AfterNower
-	Exiter    func(int)
-	KeyStore  *KeyStore
-	TxManager *TxManager
-	sigs      chan os.Signal
+	Config      Config
+	Clock       AfterNower
+	Exiter      func(int)
+	KeyStore    *KeyStore
+	TxManager   *TxManager
+	HeadTracker *HeadTracker
+	sigs        chan os.Signal
 }
 
 // NewStore will create a new database file at the config's RootDir if
@@ -37,12 +41,19 @@ func NewStore(config Config) *Store {
 		logger.Fatal(err)
 	}
 	keyStore := NewKeyStore(config.KeysDir())
+
+	ht, err := NewHeadTracker(orm)
+	if err != nil {
+		logger.Fatal(err)
+	}
+
 	store := &Store{
-		ORM:      orm,
-		Config:   config,
-		KeyStore: keyStore,
-		Exiter:   os.Exit,
-		Clock:    Clock{},
+		ORM:         orm,
+		Config:      config,
+		KeyStore:    keyStore,
+		Exiter:      os.Exit,
+		Clock:       Clock{},
+		HeadTracker: ht,
 		TxManager: &TxManager{
 			Config:    config,
 			EthClient: &EthClient{ethrpc},
@@ -84,4 +95,48 @@ func (Clock) Now() time.Time {
 // After returns the current time if the given duration has elapsed.
 func (Clock) After(d time.Duration) <-chan time.Time {
 	return time.After(d)
+}
+
+// Holds and stores the latest block header experienced by this particular node
+// in a thread safe manner. Reconstitutes the last block header from the data
+// store on reboot.
+type HeadTracker struct {
+	orm         *models.ORM
+	blockHeader *models.BlockHeader
+	mutex       sync.Mutex
+}
+
+func (ht *HeadTracker) Save(bh *models.BlockHeader) error {
+	if bh == nil {
+		return errors.New("Cannot save a nil block header")
+	}
+
+	ht.mutex.Lock()
+	if ht.blockHeader == nil || ht.blockHeader.Number.ToInt().Cmp(bh.Number.ToInt()) < 0 {
+		copy := *bh
+		ht.blockHeader = &copy
+	}
+	ht.mutex.Unlock()
+	return ht.orm.Save(bh)
+}
+
+func (ht *HeadTracker) Get() *models.BlockHeader {
+	ht.mutex.Lock()
+	defer ht.mutex.Unlock()
+	return ht.blockHeader
+}
+
+// Instantiates a new HeadTracker using the orm to persist
+// new BlockHeaders
+func NewHeadTracker(orm *models.ORM) (*HeadTracker, error) {
+	ht := &HeadTracker{orm: orm}
+	blockHeaders := []models.BlockHeader{}
+	err := orm.AllByIndex("Number", &blockHeaders, storm.Limit(1), storm.Reverse())
+	if err != nil {
+		return nil, err
+	}
+	if len(blockHeaders) > 0 {
+		ht.blockHeader = &blockHeaders[0]
+	}
+	return ht, nil
 }
