+++ utils/utils.go
@@ -208,6 +208,13 @@ func HexToString(hex string) (string, error) {
 	return string(b), err
 }
 
+func ToFilterQueryFor(fromBlock *big.Int, addresses []common.Address) ethereum.FilterQuery {
+	return ethereum.FilterQuery{
+		FromBlock: fromBlock,
+		Addresses: WithoutZeroAddresses(addresses),
+	}
+}
+
 // https://github.com/ethereum/go-ethereum/blob/762f3a48a00da02fe58063cb6ce8dc2d08821f15/ethclient/ethclient.go#L363
 func ToFilterArg(q ethereum.FilterQuery) interface{} {
 	arg := map[string]interface{}{
